WIP(login): refactor username screen
